fix: 修复音频卡顿与视频丢帧根因 — 环形缓冲区分块式、帧队列三索引、PacketQueue溢出修复
- AudioRingBuffer: 连续环形缓冲区重构为分块式(chunk-based)，每个chunk独立PTS追踪
- VideoFrameQueue: 分离decode/render/display三索引，帧级serial校验
- PacketQueue: m_totalDurationUs int→int64_t 防止2.1秒溢出导致误丢帧
- Pipeline: shutdown不再递增serial，消除use-after-free窗口
- 解码器: thread_count改为0让FFmpeg自动选择
- 系统: timeBeginPeriod(1)提升Windows定时器精度，音频默认启用
- 删除per-packet LOG_INFO，消除I/O竞争

+++ src/PacketQueue.h
@@ -47,5 +47,5 @@ private:
     std::condition_variable m_cond;
     std::atomic<bool>       m_abort{false};
     std::atomic<bool>       m_keyFrameDropped{false};
-    int                     m_totalDurationUs = 0;
+    int64_t                 m_totalDurationUs = 0;
 };
